Fixing simple issues with a build fail. Renaming util/sous_qa_setup
