+++ bin/ci-setup
@@ -23,7 +23,7 @@ sudo mv docker-compose /usr/local/bin
 
 echo Installing local tools
 go get github.com/kardianos/govendor github.com/nyarly/engulf github.com/nyarly/legendary
-go install ./util/sous_qa_setup
+go install ./dev_support/sous_qa_setup
 (cd integration/test-registry && docker-compose pull)
 
 echo Results:
